extra/streamtasksui/requirements.txt
briefcase
pillow
toml